WIP - refactoring for testing
We need to have a modular, interface-based architecture
for the conversations and in the end for combat etc.

+++ kftw2/core/src/main/kotlin/systems/RenderMapSystem.kt
@@ -4,6 +4,7 @@ import com.badlogic.ashley.core.EntitySystem
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
+import com.badlogic.gdx.math.MathUtils
 import com.badlogic.gdx.math.Vector2
 import com.badlogic.gdx.math.Vector3
 import com.badlogic.gdx.utils.PerformanceCounters
@@ -40,8 +41,8 @@ class RenderMapSystem(val fogOfWar:Boolean = false) : EntitySystem(0) {
 
       for ((x, rows) in tilesToRender.withIndex())
         for ((y, tileInstance) in rows.withIndex()) {
-          val xPos = tileInstance.x * 8f
-          val yPos = tileInstance.y * 8f
+          val xPos = (tileInstance.x * 8).toFloat()
+          val yPos = (tileInstance.y * 8).toFloat()
 
           tileInstance.baseSprite.setPosition(xPos, yPos)
           tileInstance.baseSprite.draw(batch)
